+++ src/osgui/osx/ostext.m
@@ -33,6 +33,17 @@
 
 /*---------------------------------------------------------------------------*/
 
+#if defined(MAC_OS_X_VERSION_10_14) && MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_14
+#else
+@interface OSXScrollView : NSScrollView
+{
+    uint32_t dummy;
+}
+@end
+#endif
+
+/*---------------------------------------------------------------------------*/
+
 @interface OSXTextView : NSTextView
 {
   @public
@@ -47,7 +58,8 @@
     NSMutableDictionary *dict;
     NSRange select;
     uint32_t num_chars;
-    BOOL has_focus;
+    BOOL focused;
+    BOOL focus_draw;
     BOOL is_editable;
     BOOL is_opaque;
     BOOL is_wrap_mode;
@@ -58,18 +70,43 @@
 
 /*---------------------------------------------------------------------------*/
 
+#if defined(MAC_OS_X_VERSION_10_14) && MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_14
+#else
+@implementation OSXScrollView
+
+/* Draw focus ring in older mac OSX */
+- (void)drawRect:(NSRect)rect
+{
+    OSXTextView *view = [self documentView];
+    [super drawRect:rect];
+
+    if (view->focused == YES)
+    {
+        /* Avoid draw focus ring over scrollbars */
+        if (view->focus_draw == YES)
+        {
+            NSRect r = [self bounds];
+            r.origin.x += 1;
+            r.origin.y += 1;
+            r.size.width -= 2;
+            r.size.height -= 2;
+            NSSetFocusRingStyle(NSFocusRingOnly);
+            NSRectFill(r);
+            view->focus_draw = NO;
+        }
+    }
+}
+
+@end
+#endif
+
+/*---------------------------------------------------------------------------*/
+
 @implementation OSXTextView
 
 - (void)dealloc
 {
     [super dealloc];
-}
-
-/*---------------------------------------------------------------------------*/
-
-- (void)drawRect:(NSRect)rect
-{
-    [super drawRect:rect];
 }
 
 /*---------------------------------------------------------------------------*/
@@ -149,6 +186,7 @@ static void i_replace_seltext(OSXTextView *lview, NSRange range, const char_t *t
 @interface OSXTextViewDelegate : NSObject
 #endif
 {
+    uint32_t dummy;
 }
 
 @end
@@ -272,9 +310,14 @@ OSText *ostext_create(const uint32_t flags)
     heap_auditor_add("OSXTextView");
     view->is_editable = NO;
     view->is_opaque = YES;
-    view->has_focus = NO;
+    view->focused = NO;
+    view->focus_draw = NO;
     view->is_wrap_mode = YES;
+#if defined(MAC_OS_X_VERSION_10_14) && MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_14
     view->scroll = [[NSScrollView alloc] initWithFrame:NSZeroRect];
+#else
+    view->scroll = [[OSXScrollView alloc] initWithFrame:NSZeroRect];
+#endif
     view->dict = [[NSMutableDictionary alloc] init];
     view->ffamily[0] = '\0';
     view->fsize = REAL32_MAX;
@@ -312,7 +355,7 @@ OSText *ostext_create(const uint32_t flags)
 
     /* OSText allways have border */
     {
-        [view setFocusRingType:NSFocusRingTypeExterior];
+        [view->scroll setFocusRingType:NSFocusRingTypeExterior];
     }
 
     return (OSText *)view->scroll;
@@ -637,7 +680,7 @@ void ostext_property(OSText *view, const gui_text_t param, const void *value)
             lview->select = NSMakeRange(start, end - start);
         }
 
-        if (lview->has_focus)
+        if (lview->focused)
         {
             [lview setSelectedRange:lview->select];
         }
@@ -792,7 +835,8 @@ void ostext_focus(OSText *view, const bool_t focus)
     cassert_no_null(view);
     lview = [(NSScrollView *)view documentView];
     cassert_no_null(lview);
-    lview->has_focus = (BOOL)focus;
+    lview->focused = (BOOL)focus;
+    lview->focus_draw = YES;
 
     if (lview->OnFocus != NULL)
     {
@@ -809,6 +853,11 @@ void ostext_focus(OSText *view, const bool_t focus)
         lview->select = [lview selectedRange];
         [lview setSelectedRange:NSMakeRange(0, 0)];
     }
+
+#if defined(MAC_OS_X_VERSION_10_14) && MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_14
+#else
+    [lview->scroll setNeedsDisplay:YES];
+#endif
 }
 
 /*---------------------------------------------------------------------------*/